The correct audio system is used depending on whether the file is available
