+++ source/sounds.cpp
@@ -11,6 +11,79 @@
 int sounds_pos = 0;
 sound_t loaded_sounds[8] = {SOUND_NONE}; // circular buffer (must be power of two)
 music_t loaded_music = MUSIC_NONE;
+bool streamOpen = false;
+
+FILE *file;
+
+void stopStream()
+{
+	if (!streamOpen)
+		return;
+	mmStreamClose();
+	fclose(file);
+	streamOpen = false;
+}
+
+bool streamIsOpen()
+{
+	return streamOpen;
+}
+
+bool songIsPlaying()
+{
+	return loaded_music != MUSIC_NONE;
+}
+
+mm_word stream(mm_word length, mm_addr dest, mm_stream_formats format)
+{
+	if (file)
+	{
+		size_t samplesize = 1;
+		switch (format)
+		{
+			case MM_STREAM_8BIT_MONO: samplesize = 1;
+				break;
+			case MM_STREAM_8BIT_STEREO: samplesize = 2;
+				break;
+			case MM_STREAM_16BIT_MONO: samplesize = 2;
+				break;
+			case MM_STREAM_16BIT_STEREO: samplesize = 4;
+				break;
+		}
+
+		int res = fread(dest, samplesize, length, file);
+		if (res)
+		{
+			length = res;
+		}
+		else
+		{
+			mmStreamClose();
+			fclose(file);
+			length = 0;
+			streamOpen = false;
+		}
+	}
+	return length;
+}
+
+void playStreamSong()
+{
+	if (streamOpen)
+		return;
+	mm_stream mystream;
+	mystream.buffer_length = 1024;
+	mystream.callback = stream;
+	mystream.timer = MM_TIMER1;
+	mystream.manual = true;
+
+	mystream.sampling_rate = 22050;
+	mystream.format = MM_STREAM_16BIT_STEREO;
+	mmStreamOpen(&mystream);
+	mmStreamUpdate();
+	mmStreamUpdate();
+	streamOpen = true;
+}
 
 void initSound(void)
 {
@@ -19,7 +92,7 @@ void initSound(void)
 
 static void loadSound(sound_t sound)
 {
-	/*unsigned int i;
+	unsigned int i;
 
 	if (sound == SOUND_NONE)
 		return;
@@ -34,12 +107,12 @@ static void loadSound(sound_t sound)
 		mmUnloadEffect(loaded_sounds[sounds_pos]);
 	loaded_sounds[sounds_pos] = sound;
 	sounds_pos = (sounds_pos + 1) & (LENGTH(loaded_sounds) - 1);
-	mmLoadEffect(sound);*/
+	mmLoadEffect(sound);
 }
 
 void playSound(sound_t sound)
 {
-	/*if (sound == SOUND_NONE)
+	if (sound == SOUND_NONE)
 		return;
 
 	if (isWifi())
@@ -52,13 +125,13 @@ void playSound(sound_t sound)
 	}
 
 	loadSound(sound);
-	mmEffect(sound);*/
+	mmEffect(sound);
 }
 
 void playSoundNifi(sound_t sound)
 {
-	/*loadSound(sound);
-	mmEffect(sound);*/
+	loadSound(sound);
+	mmEffect(sound);
 }
 
 void playMusic(music_t music)
@@ -66,13 +139,19 @@ void playMusic(music_t music)
 	if (music == MUSIC_NONE)
 		return;
 
-	/*if (music != loaded_music)
+	if (music != loaded_music)
 	{
 		stopMusic();
-		mmLoad(music);
+		bool canStream = (file = fopen("MineCraft Remix - Calm 3 Chillstep.raw", "rb")) != NULL;
+		if (!canStream)
+		{
+			mmLoad(music);
+			mmStart(music, MM_PLAY_LOOP); //Prevents music restarting
+		}
+		else
+			playStreamSong();
 		loaded_music = music;
-		mmStart(music, MM_PLAY_LOOP); //Prevents music restarting
-	}*/
+	}
 }
 
 void stopMusic(void)
@@ -80,7 +159,12 @@ void stopMusic(void)
 	if (loaded_music == MUSIC_NONE)
 		return;
 
-	/*mmStop();
-	mmUnload(loaded_music);
-	loaded_music = MUSIC_NONE;*/
+	if (!streamOpen)
+	{
+		mmStop();
+		mmUnload(loaded_music);
+	}
+	else
+		stopStream();
+	loaded_music = MUSIC_NONE;
 }
